improve gen
